pyproject.toml
[build-system]
requires = ["maturin>=1.9,<2.0"]
build-backend = "maturin"

[project]
name = "django-bolt"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = [
    "Django>=4.2",
    "click>=8.1",
    "pydantic>=2",
    "msgspec>=0.18",
    "maturin>=1.9.4",
]
[project.scripts]
django-bolt = "django_bolt.cli:main"
[tool.maturin]
features = ["pyo3/extension-module"]
module-name = "django_bolt._core"
python-source = "python"

[tool.uv.workspace]
members = [
    "ignore",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pyjwt>=2.10.1",
    "pytest-asyncio>=1.2.0",
]
